+++ docs/changelog.md
@@ -6,6 +6,13 @@ nav_order: 2
 
 # Changelog
 
+# Unreleased changes
+
+### Internals
+
+- Return `bytes` from `BaseAddress.to_knx()` instead of `tuple[int, int]`. This is used in `IndividualAddress` and `GroupAddress`.
+- Add `BaseAddress.from_knx()` to instantiate from `bytes`, remove instantiation form `tuple[int, int]`.
+
 # 2.3.0 Routing security, DPTs and CEMI-Refactoring 2023-01-10
 
 ### DPTs